Duplicate But Just To Be Sure
Using the submit to source control on UE will commit to the repository
automatically but as far as GitHub knows the changes are still there. So
just to be safe I will commit this now but this is just an FYI to
anybody so confusion can be avoided. Sorry again for any duplicates.