Battle and Level - LOTS OF CHANGES
I keep forgetting to sync my commits to git so I keep losing progress
but I got in the habit of backing up stuff so I just had to adapt some
things. So I did a bunch including:
1. Adjusted Chris's lights, cause I can't see jack diddly
2. Introducing the almighty TileMap: takes a tileset, and turns it into
a grid. I added a collision layer for the floor, and a background layer.
Level design just got way easier for us.
3. The AbyssGameInstance: Persistent data storage across the game. Acts
as a data manager to send info back and forth across our levels and
battle.
3. General Enemy Struct: Data structure that contains a bunch of info
for enemies when we go into battle. A template for sending info to the
battle, including HUD pics and enemy stats.
4. Battle transitions: There is a Fort Battle blueprint or something
like that that is our collision zone for triggering battles. These
contain an array of possible enemies, that are randomly selected, then
stored in the GameInstance, then sent to the Battle, where the HUD and
the enemies will change to reflect this data.

NOTE: None of these are FULLY IMPLEMENTED yet, so it will be relatively
broken! Working on them a lot today, so will update as I go.