Get Player Location (Maybe)
When the battle scene starts, the player location is displayed in the
upper left (since I wasn't sure what to do with the value just yet).
The values are close enough to the character's location values just
before hitting the battle trigger, so I'm pretty sure it's the right
value.